Update cfp.md
Removed this wording at Preslav Nakov's request:

The "notification (or withdrawal)" phrase used in the ACL guidelines refers to notification of acceptance/rejection from a *ACL conference (or workshop). This means that ARR submissions must remain anonymized until a conference or workshop has accepted or rejected, or until the submission is withdrawn.